refactor: remove unnecessary exectuors
