+++ ams/server/src/main/java/com/netease/arctic/server/optimizing/OptimizingQueue.java
@@ -18,7 +18,6 @@
 
 package com.netease.arctic.server.optimizing;
 
-import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import com.netease.arctic.AmoroTable;
 import com.netease.arctic.api.OptimizerProperties;
 import com.netease.arctic.api.OptimizingTaskId;
@@ -38,7 +37,6 @@ import com.netease.arctic.server.table.ServerTableIdentifier;
 import com.netease.arctic.server.table.TableManager;
 import com.netease.arctic.server.table.TableRuntime;
 import com.netease.arctic.server.table.TableRuntimeMeta;
-import com.netease.arctic.server.utils.TimeoutUtils;
 import com.netease.arctic.table.ArcticTable;
 import com.netease.arctic.utils.ArcticDataFiles;
 import com.netease.arctic.utils.CompatiblePropertyUtil;
@@ -64,8 +62,6 @@ import java.util.Queue;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executor;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.concurrent.LinkedTransferQueue;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.locks.Condition;
@@ -91,7 +87,6 @@ public class OptimizingQueue extends PersistentBase {
   private final int maxPlanningParallelism;
   private final OptimizerGroupMetrics metrics;
   private ResourceGroup optimizerGroup;
-  private final ExecutorService asyncPersistExecutors;
 
   public OptimizingQueue(
       TableManager tableManager,
@@ -112,11 +107,6 @@ public class OptimizingQueue extends PersistentBase {
             optimizerGroup.getName(), MetricManager.getInstance().getGlobalRegistry(), this);
     this.metrics.register();
     tableRuntimeMetaList.forEach(this::initTableRuntime);
-    this.asyncPersistExecutors =
-        Executors.newCachedThreadPool(
-            new ThreadFactoryBuilder()
-                .setNameFormat("async-persist-executors-" + optimizerGroup.getName())
-                .build());
   }
 
   private void initTableRuntime(TableRuntimeMeta tableRuntimeMeta) {
@@ -640,9 +630,7 @@ public class OptimizingQueue extends PersistentBase {
                 taskMap
                     .values()
                     .forEach(
-                        task ->
-                            TimeoutUtils.runWithTimeout(
-                                task::tryCanceling, asyncPersistExecutors, 10, TimeUnit.SECONDS)),
+                        task -> task.tryCanceling(5, TimeUnit.SECONDS)),
             () ->
                 doAs(
                     OptimizingMapper.class,
